@@ -40,54 +43,57 @@
   },
   "homepage": "https://github.com/angular/angularfire#readme",
   "dependencies": {
-    "@angular-devkit/architect": ">= 0.900 < 0.1200",
-    "@angular-devkit/core": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular-devkit/schematics": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/common": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/compiler": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/core": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/platform-browser": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/platform-browser-dynamic": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/router": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "firebase": "^7.0 || ^8.0",
+    "@angular-devkit/architect": "^0.1200.0",
+    "@angular-devkit/core": "^12.0.0",
+    "@angular-devkit/schematics": "^12.0.0",
+    "@angular/common": "^12.0.0",
+    "@angular/compiler": "^12.0.0",
+    "@angular/core": "^12.0.0",
+    "@angular/platform-browser": "^12.0.0",
+    "@angular/platform-browser-dynamic": "^12.0.0",
+    "@angular/router": "^12.0.0",
+    "firebase": "^9.0.0",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.6.0",
-    "firebase-tools": "^8.0.0",
+    "firebase-tools": "^9.0.0",
     "fs-extra": "^8.0.1",
     "fuzzy": "^0.1.3",
     "husky": "^4.2.5",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.0.1",
+    "jsonc-parser": "^3.0.0",
     "open": "^7.0.3",
-    "rxfire": "^3.9.7 || ^4.0.0",
-    "rxjs": "^6.5.3",
+    "rxfire": "^6.0.0",
+    "rxjs": "~6.6.0",
     "semver": "^7.1.3",
-    "tslib": "^1.10.0 || ^2.0.0",
+    "tslib": "^2.1.0",
+    "webpack": "^5.35.0",
     "ws": "^7.2.1",
     "xhr2": "^0.1.4",
-    "zone.js": "0.10.2"
+    "zone.js": "~0.11.4"
   },
   "optionalDependencies": {
     "bufferutil": "^4.0.1",
     "utf-8-validate": "^5.0.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": ">= 0.900 < 0.1200",
-    "@angular-devkit/build-ng-packagr": ">= 0.900 < 0.1200",
-    "@angular/animations": " ^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/cli": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/compiler-cli": "^9.0.0 || ^10.0.0 || ^11.0.0",
-    "@angular/platform-server": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular-devkit/build-angular": "^12.0.0",
+    "@angular/animations": "~12.0.0",
+    "@angular/cli": "^12.0.0",
+    "@angular/compiler-cli": "^12.0.0",
+    "@angular/platform-server": "^12.0.0",
     "@types/fs-extra": "^7.0.0",
     "@types/gzip-size": "^5.1.1",
     "@types/inquirer": "^0.0.44",
     "@types/jasmine": "^3.3.13",
+    "@types/lodash.isequal": "^4.5.5",
     "@types/node": "^12.6.2 < 12.12.42",
     "@types/request": "0.0.30",
     "@types/semver": "^7.1.0",
     "codelyzer": "^6.0.0",
     "concurrently": "^2.2.0",
     "conventional-changelog-cli": "^1.2.0",
+    "file-loader": "^6.2.0",
     "firebase-functions-test": "^0.2.2",
     "globalthis": "^1.0.1",
     "gzip-size": "^5.1.1",
@@ -117,10 +125,10 @@
     "tslint": "~6.1.0",
     "ttypescript": "^1.5.12",
     "typedoc": "^0.16.4",
-    "typescript": ">=4.0.0 <4.1.0"
+    "typescript": ">=4.2.3 <4.3.0"
   },
   "typings": "index.d.ts",
   "resolutions": {
-    "webpack": "5.4.0"
+    "webpack": "^5.35.0"
   }
 }
