@@ -3,72 +3,76 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "start:emulated": "concurrently -n ng,firebase -c red,yellow \"ng serve -c emulated\" \"firebase emulators:start\"",
-    "build": "ng build",
+    "start": "ng serve --hmr",
+    "build": "ng build --stats-json",
+    "build:prod": "ng build --configuration production --stats-json",
+    "analyze": "webpack-bundle-analyzer dist/sample/browser/stats.json",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "lint": "ng lint",
-    "deploy": "ng deploy",
     "dev:ssr": "ng run sample:serve-ssr",
-    "bust": "rm -rf node_modules/@angular/fire && yarn cache clean @angular/fire && yarn add ../dist/packages-dist",
     "serve:ssr": "node dist/sample/server/main.js",
-    "build:ssr": "ng build --prod && ng run sample:server:production",
+    "build:ssr": "ng build && ng run sample:server",
     "prerender": "ng run sample:prerender"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.0",
-    "@angular/common": "~11.0.0",
-    "@angular/compiler": "~11.0.0",
-    "@angular/core": "~11.0.0",
+    "@angular/animations": "^12.0.0",
+    "@angular/common": "^12.0.0",
+    "@angular/compiler": "^12.0.0",
+    "@angular/core": "^12.0.0",
     "@angular/fire": "../dist/packages-dist",
-    "@angular/forms": "~11.0.0",
-    "@angular/platform-browser": "~11.0.0",
-    "@angular/platform-browser-dynamic": "~11.0.0",
-    "@angular/platform-server": "~11.0.0",
-    "@angular/router": "~11.0.0",
-    "@angular/service-worker": "^11.0.0",
-    "@nguniversal/express-engine": "~10.1.0",
-    "core-js": "^3.6.5",
-    "firebase": "^8.0.0",
-    "first-input-delay": "^0.1.3",
-    "proxy-polyfill": "^0.3.2",
-    "rxjs": "~6.6.3",
-    "tslib": "^2.0.1",
-    "whatwg-fetch": "^3.4.1",
-    "zone.js": "0.10.3"
+    "@angular/forms": "^12.0.0",
+    "@angular/platform-browser": "^12.0.0",
+    "@angular/platform-browser-dynamic": "^12.0.0",
+    "@angular/platform-server": "^12.0.0",
+    "@angular/router": "^12.0.0",
+    "@angular/service-worker": "^12.0.0",
+    "@nguniversal/express-engine": "^12.1.0",
+    "cross-fetch": "^3.1.4",
+    "express": "^4.15.2",
+    "firebase": "^9.0.0",
+    "lodash.isequal": "^4.5.0",
+    "rxfire": "^6.0.0",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.1.0",
+    "ws": "^8.2.0",
+    "xhr2": "^0.2.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.1001.3",
-    "@angular-devkit/build-angular": "~0.1100.0",
-    "@angular/cli": "~11.0.0",
-    "@angular/compiler-cli": "~11.0.0",
-    "@angular/language-service": "~11.0.0",
-    "@firebase/app-types": "^0.6.1",
-    "@nguniversal/builders": "^10.1.0",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^6.0.0",
-    "concurrently": "^5.3.0",
-    "dir-loader": "^0.3.0",
-    "express": "^4.17.1",
-    "express-serve-static-core": "^0.1.1",
-    "firebase-admin": "^8.13.0",
-    "firebase-functions": "^3.11.0",
+    "@angular-devkit/architect": "^0.1200.0",
+    "@angular-devkit/build-angular": "^12.0.0",
+    "@angular/cli": "^12.0.0",
+    "@angular/compiler-cli": "^12.0.0",
+    "@nguniversal/builders": "^12.1.0",
+    "@types/express": "^4.17.0",
+    "@types/jasmine": "~3.6.0",
+    "@types/node": "^12.11.1",
+    "file-loader": "^6.2.0",
+    "firebase-admin": "^9.11.1",
+    "firebase-functions": "^3.6.0",
     "firebase-functions-test": "^0.2.2",
-    "firebase-tools": "^8.16.1",
+    "firebase-tools": "^9.0.0",
     "fuzzy": "^0.1.3",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.0.1",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.1",
+    "jasmine-core": "~3.7.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "jsonc-parser": "^3.0.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ts-node": "~9.0.0",
+    "ng-packagr": "^12.0.0",
+    "open": "^7.0.3",
+    "ts-node": "~9.1.1",
     "tslint": "~6.1.3",
-    "typescript": "~4.0.5"
+    "typescript": "~4.2.3",
+    "webpack-bundle-analyzer": "^4.4.1"
+  },
+  "resolutions": {
+    "webpack": "^5.35.0"
   }
 }
