@@ -1,33 +1,34 @@
 import { Component, OnInit } from '@angular/core';
-import { AngularFirestore } from '@angular/fire/firestore';
-import { Observable } from 'rxjs';
-import { startWith, tap } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
+import { startWith, switchMap, tap } from 'rxjs/operators';
 import { makeStateKey, TransferState } from '@angular/platform-browser';
-import { trace } from '@angular/fire/performance';
+import { traceUntilFirst } from '@angular/fire/performance';
 
 @Component({
   selector: 'app-firestore',
   template: `<p>
     Firestore!
     {{ testDocValue$ | async | json }}
-    {{ persistenceEnabled$ | async }}
+    {{ persistenceEnabled | async }}
   </p>`,
   styles: [``]
 })
 export class FirestoreComponent implements OnInit {
 
-  public readonly persistenceEnabled$: Observable<boolean>;
   public readonly testDocValue$: Observable<any>;
+  public persistenceEnabled: Promise<boolean> = Promise.resolve(false);
 
-  constructor(state: TransferState, firestore: AngularFirestore) {
-    const doc = firestore.doc('test/1');
-    const key = makeStateKey(doc.ref.path);
+  constructor(state: TransferState) {
+    const key = makeStateKey<unknown>('FIRESTORE');
     const existing = state.get(key, undefined);
-    this.testDocValue$ = firestore.doc('test/1').valueChanges().pipe(
-      trace('firestore'),
-      existing ? startWith(existing) : tap(it => state.set(key, it))
+    this.testDocValue$ = of(existing).pipe(
+      switchMap(() => import('./lazyFirestore')),
+      tap(({ persistenceEnabled }) => this.persistenceEnabled = persistenceEnabled),
+      switchMap(({ valueChanges }) => valueChanges),
+      traceUntilFirst('firestore'),
+      tap(it => state.set(key, it)),
+      existing ? startWith(existing) : tap(),
     );
-    this.persistenceEnabled$ = firestore.persistenceEnabled$;
   }
 
   ngOnInit(): void {
