@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { AngularFireRemoteConfig, mapToObject } from '@angular/fire/remote-config';
-import { trace } from '@angular/fire/performance';
+import { RemoteConfig, getAllChanges } from '@angular/fire/remote-config';
+import { traceUntilFirst } from '@angular/fire/performance';
 import { Observable } from 'rxjs';
 
 @Component({
@@ -17,8 +17,8 @@ export class RemoteConfigComponent implements OnInit {
 
   readonly change$: Observable<any>;
 
-  constructor(public readonly remoteConfig: AngularFireRemoteConfig) {
-    this.change$ = remoteConfig.parameters.pipe(trace('remote-config'), mapToObject({}));
+  constructor(public readonly remoteConfig: RemoteConfig) {
+    this.change$ = getAllChanges(this.remoteConfig).pipe(traceUntilFirst('remote-config'));
   }
 
   ngOnInit(): void {
