@@ -96,9 +96,8 @@ const getPackageJson = (context: BuilderContext, workspaceRoot: string, options:
     const angularJson = JSON.parse(readFileSync(join(workspaceRoot, 'angular.json')).toString());
     // tslint:disable-next-line:no-non-null-assertion
     const server = angularJson.projects[context.target!.project].architect.server;
-    const serverOptions = server && server.options;
-    const externalDependencies = serverOptions && serverOptions.externalDependencies || [];
-    const bundleDependencies = serverOptions && serverOptions.bundleDependencies;
+    const externalDependencies = server?.options?.externalDependencies || [];
+    const bundleDependencies = server?.options?.bundleDependencies ?? true;
     if (bundleDependencies !== true) {
       if (existsSync(join(workspaceRoot, 'package.json'))) {
         const packageJson = JSON.parse(readFileSync(join(workspaceRoot, 'package.json')).toString());
@@ -172,10 +171,12 @@ export const deployToFunction = async (
     defaultFunction(serverOut, options)
   );
 
-  fsHost.renameSync(
-    join(newClientPath, 'index.html'),
-    join(newClientPath, 'index.original.html')
-  );
+  try {
+    fsHost.renameSync(
+      join(newClientPath, 'index.html'),
+      join(newClientPath, 'index.original.html')
+    );
+  } catch (e) { }
 
   if (options.preview) {
     const port = 5000; // TODO make this configurable
@@ -254,16 +255,16 @@ export default async function deploy(
     const winston = require('winston');
     const tripleBeam = require('triple-beam');
 
-    firebaseTools.logger.add(
-      new winston.transports.Console({
-        level: 'info',
-        format: winston.format.printf((info) =>
-          [info.message, ...(info[tripleBeam.SPLAT] || [])]
-            .filter((chunk) => typeof chunk === 'string')
-            .join(' ')
-        )
-      })
-    );
+    const logger = new winston.transports.Console({
+      level: 'info',
+      format: winston.format.printf((info) =>
+        [info.message, ...(info[tripleBeam.SPLAT] || [])]
+          .filter((chunk) => typeof chunk === 'string')
+          .join(' ')
+      )
+    });
+
+    firebaseTools.logger.logger.add(logger);
 
     if (serverBuildTarget) {
       await deployToFunction(
